Updated README and directory names.
