+++ bin/autotask
@@ -18,7 +18,7 @@ if (file_exists(getenv('AUTOTASK_GENERATOR_DIRECTORY') . '/.env')) {
 
 // Setup our console app
 $application = new Symfony\Component\Console\Application();
-$application->add( new Anteris\Autotask\Generator\Command\MakeEndpointCommand() );
-$application->add( new Anteris\Autotask\Generator\Command\MakeDefaultsCommand() );
-$application->add( new Anteris\Autotask\Generator\Command\MakeSupportFilesCommand() );
+$application->add( new Anteris\Autotask\Generator\Commands\MakeEndpointCommand() );
+$application->add( new Anteris\Autotask\Generator\Commands\MakeDefaultsCommand() );
+$application->add( new Anteris\Autotask\Generator\Commands\MakeSupportFilesCommand() );
 $application->run();
