+++ src/Commands/MakeDefaultsCommand.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Anteris\Autotask\Generator\Command;
+namespace Anteris\Autotask\Generator\Commands;
 
+use Anteris\Autotask\Generator\Helpers\Api;
 use Illuminate\Support\Env;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class MakeEndpointCommand extends AbstractMakeCommand
+class MakeDefaultsCommand extends AbstractMakeCommand
 {
     /** @var string The name of this command.  */
-    protected static $defaultName = 'make:endpoint';
+    protected static $defaultName = 'make:defaults';
 
     /**
      * Configures the command information.
@@ -21,11 +21,11 @@ class MakeEndpointCommand extends AbstractMakeCommand
     protected function configure()
     {
         $this
-            ->setDescription('Creates a new PHP class for interacting with an Autotask service.')
-            ->addArgument('entity', InputArgument::REQUIRED, 'The entity to generate classes for.')
+            ->setDescription('Creates a new PHP class for the default Autotask services.')
             ->addOption('no-cache', null, InputOption::VALUE_NONE, 'Specifies whether any previous cached files should be ignored (still makes use of cache for efficiency, but clears afterward).')
-            ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Sets the output directory for the generated class.')
-            ->addOption('force', 'f', InputOption::VALUE_NONE, 'Overwrites previous classes if they exist.');
+            ->addOption('force', 'f', InputOption::VALUE_NONE, 'Overwrites previous classes if they exist.')
+            ->addOption('output', 'o', InputOption::VALUE_REQUIRED, 'Sets the output directory for the generated classes.');
+
     }
 
     /**
@@ -34,10 +34,13 @@ class MakeEndpointCommand extends AbstractMakeCommand
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->setupEnvironment($input, $output);
+        $endpoints = Api::endpoints();
 
         try {
-            $output->writeln('Generating classes for ' . $input->getArgument('entity'));
-            $this->generator->makeResource($input->getArgument('entity'));
+            foreach ($endpoints as $service) {
+                $output->writeln('Generating classes for ' . $service);
+                $this->generator->makeResource($service);
+            }
         } catch (\Exception $error) {
             $output->writeln(
                 '<error>There was an error creating that endpoint: ' .
@@ -48,8 +51,10 @@ class MakeEndpointCommand extends AbstractMakeCommand
         }
 
         $output->writeln(
-            '<info>Successfully created endpoint in ' .
-            (microtime(true) - $_SERVER['REQUEST_TIME_FLOAT']) .
+            '<info>Successfully created '.
+            count($endpoints).
+            ' endpoints in '.
+            (microtime(true) - $_SERVER['REQUEST_TIME_FLOAT']).
             ' seconds!</info>'
         );
         $output->writeln('');
