+++ src/Helpers/Str.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Anteris\Autotask\Generator\Helper;
+namespace Anteris\Autotask\Generator\Helpers;
 
 use Illuminate\Support\Str as StrHelper;
